Move clean fixtures to conftest
