+++ tests/integration/azure/test_locations.py
@@ -257,29 +257,6 @@ def test_run_validate_acl(make_cluster_permissions, ws, make_user, make_cluster,
         )
 
 
-def delete_ucx_created_resources(api_client):
-    """Delete UCX created resources"""
-    for resource in api_client.list():
-        if resource.name is not None and resource.comment is not None and resource.comment == "Created by UCX":
-            api_client.delete(resource.name, force=True)
-
-
-@pytest.fixture
-def clean_storage_credentials(az_cli_ctx):
-    """Clean test generated storage credentials."""
-    delete_ucx_created_resources(az_cli_ctx.workspace_client.storage_credentials)
-    yield
-    delete_ucx_created_resources(az_cli_ctx.workspace_client.storage_credentials)
-
-
-@pytest.fixture
-def clean_external_locations(az_cli_ctx):
-    """Clean test generated external locations."""
-    delete_ucx_created_resources(az_cli_ctx.workspace_client.external_locations)
-    yield
-    delete_ucx_created_resources(az_cli_ctx.workspace_client.external_locations)
-
-
 def test_run_external_locations_using_access_connector(
     clean_storage_credentials,
     clean_external_locations,
